updated WindowsMusicHelper.exe (updated to be 5.4mb tiny like I wanted and be slightly better) + updated musicoverlay to be better. took forever because thinking is painful recently.
TODO: restart exe on error maybe?
